orchestrator/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
flask_cors
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.5.0
pymongo==3.10.1
six==1.14.0
Werkzeug==0.14.1
Click==7.0
dnspython==1.16.0
Flask==1.0.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.5.0
pymongo==3.10.1
six==1.14.0
Werkzeug==0.14.1
Click==7.0
dnspython==1.16.0
Flask==1.0.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.5.0
pymongo==3.10.1
six==1.14.0
Werkzeug==0.14.1
